% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rwMetropolisC}
\alias{rwMetropolisC}
\title{A random walk Metropolis sampler for generating the standard Laplace distribution using Rcpp}
\usage{
rwMetropolisC(sigma, x0, N)
}
\arguments{
\item{sigma}{The standard deviation of proposal distribution N(Xt,sigma^2)}

\item{N}{the length of the chain}

\item{X0}{initial value X0}
}
\value{
a random sample of size \code{n}
}
\description{
A random walk Metropolis sampler for generating the standard Laplace distribution using Rcpp
}
\examples{
\dontrun{
rwC <- rw.MetropolisC(.5, 25, 2000)
 plot(a, rwC$x, type = "s", main = "", xlab = "sigma = 0.5", ylab = "x")
}
}
