% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetropolisR.R
\name{rwMetropolisR}
\alias{rwMetropolisR}
\title{A random walk Metropolis sampler for generating the standard Laplace distribution using R}
\usage{
rwMetropolisR(sigma, x0, N)
}
\arguments{
\item{sigma}{The standard deviation of proposal distribution N(Xt,sigma^2)}

\item{N}{the length of the chain}

\item{X0}{initial value X0}
}
\value{
a random sample of size \code{n}
}
\description{
A random walk Metropolis sampler for generating the standard Laplace distribution using R
}
\examples{
\dontrun{
rwR <- rw.MetropolisR(.5, 25, 2000)
plot(a, rwR$x, type = "s", main = "", xlab = "sigma = 0.5", ylab = "x")
}
}
