\name{SC19023-package}
\alias{SC19023-package}
\alias{SC19023}
\docType{package}
\title{
  Compare the efficiency of R and C++
}
\description{
  The simple R package developed to compare the performance of R and R++ (implemented through the R package _Rcpp_) for the homework of 'Statistical Computing' course.
}
\details{
  Two functions are considered, namely, _linear_reg_ (Regression analysis on centralized data) and _Metropolis_ (Implement the random walk version of the Metropolis sampler to generate the target distribution N(Xt,sigma^2)). For the second function, both R and Rcpp versions are produced. Namely _MetropolisR_ for R and _MetropolisC_ for C++.
}
\author{
Wen Ma, 809450179@qq.com.

Maintainer: Wen Ma <809450179@qq.com>
}
\keyword{ package }

